refactor(gore-as): lift emit_all_tree into the library
Move the full-tree .as emission (cross-module free-function
de-collision + safe-path write loop) out of the gore CLI
AsCmd::EmitAll arm into gore_as::cache::emit_all::emit_all_tree so
the compile orchestration can reuse it. Behavior is unchanged: the
lift is byte-faithful, only param/return I/O differs (takes parsed
mods/refs + outdir, returns EmitAllStats, maps fs errors to
EmitAllError::Io). rename_free_fn moves verbatim into the lib; the
CLI arm now delegates.
