+++ crates/gore/src/cmd/as_cache.rs
@@ -1,30 +1,8 @@
-use std::collections::{HashMap, HashSet};
 use std::path::PathBuf;
 
 use anyhow::{Context, Result};
 use clap::Subcommand;
 
-/// Rename whole-word free occurrences of `name` to `newname` in `src` — occurrences NOT preceded
-/// by `.` (so member calls `obj.name(...)` are left alone; only the decl + free calls change).
-fn rename_free_fn(src: &str, name: &str, newname: &str) -> String {
-    let (b, nb) = (src.as_bytes(), name.as_bytes());
-    let word = |c: u8| c.is_ascii_alphanumeric() || c == b'_';
-    let mut out: Vec<u8> = Vec::with_capacity(b.len());
-    let mut i = 0;
-    while i < b.len() {
-        let hit = b[i..].starts_with(nb)
-            && (i == 0 || (!word(b[i - 1]) && b[i - 1] != b'.' && b[i - 1] != b':'))
-            && (i + nb.len() >= b.len() || !word(b[i + nb.len()]));
-        if hit {
-            out.extend_from_slice(newname.as_bytes());
-            i += nb.len();
-        } else {
-            out.push(b[i]);
-            i += 1;
-        }
-    }
-    String::from_utf8(out).unwrap_or_else(|_| src.to_string())
-}
 use gore_as::cache::header::CacheHeader;
 use gore_as::cache::scan::scan_strings;
 use gore_as::cache::splice::splice_auto;
@@ -208,114 +186,10 @@ pub fn run(cmd: AsCmd) -> Result<()> {
             if let Some(api) = load_native_api(&file) {
                 refs.set_native_api(api);
             }
-            // Resolve the output root up front so a symlinked `outdir` is followed to its real
-            // target ONCE here — the per-entry component check below then guards the untrusted
-            // cache paths against that real root. (create_dir_all so canonicalize succeeds.)
-            std::fs::create_dir_all(&outdir).with_context(|| format!("creating {}", outdir.display()))?;
-            let outdir = outdir.canonicalize().with_context(|| format!("resolving {}", outdir.display()))?;
-            // Cross-module free-function collisions: AngelScript compiles all loose .as into ONE
-            // global scope, so two modules each defining `Foo(<same params>)` (even with different
-            // return types) collide as "a function with the same name and parameters already
-            // exists". Find such names and rename each per-module — a function's decl and its
-            // intra-module free calls live in the same emitted file, so a file-local rename
-            // de-collides without breaking resolution (cross-module calls of these don't occur).
-            let mut sig_mods: HashMap<String, HashSet<usize>> = HashMap::new();
-            for (i, m) in mods.iter().enumerate() {
-                for f in &m.functions {
-                    // generated factory accessors are skipped at emit (not free-emitted) — never
-                    // rename them, or free CALLS to the native binding would be broken.
-                    if matches!(f.name.as_str(),
-                        "Spawn" | "Get" | "GetOrCreate" | "Create" | "GetG1R" | "StaticClass") {
-                        continue;
-                    }
-                    // precise signature (render, not base_name) so only GENUINE same-signature
-                    // collisions are flagged — a coarse match falsely flags distinct overloads and
-                    // would rename (and break) functions that are validly called cross-module.
-                    let ptys: Vec<String> = f.params.iter().map(|p| p.ty.render(&refs)).collect();
-                    sig_mods.entry(format!("{}({})", f.name, ptys.join(","))).or_default().insert(i);
-                }
-            }
-            // A name is safe to file-locally rename in a module only if EVERY emittable free
-            // function of that name in the module has a colliding signature. If the module also has
-            // a NON-colliding same-name overload, the name-based `rename_free_fn` would rewrite that
-            // overload's decl + calls too, breaking other modules that call it by its original name.
-            // In that mixed case leave the name un-renamed: the genuine collision then surfaces at
-            // generate as a "already exists" stub (rare, safe) instead of silently breaking a valid
-            // cross-module call to the non-colliding overload.
-            let colliding_sigs: HashSet<&str> = sig_mods
-                .iter()
-                .filter(|(_, modset)| modset.len() > 1)
-                .map(|(sig, _)| sig.as_str())
-                .collect();
-            let mut colliding_in: HashMap<usize, HashSet<String>> = HashMap::new();
-            for (i, m) in mods.iter().enumerate() {
-                let mut sigs_by_name: HashMap<&str, Vec<String>> = HashMap::new();
-                for f in &m.functions {
-                    if matches!(f.name.as_str(),
-                        "Spawn" | "Get" | "GetOrCreate" | "Create" | "GetG1R" | "StaticClass") {
-                        continue;
-                    }
-                    let ptys: Vec<String> = f.params.iter().map(|p| p.ty.render(&refs)).collect();
-                    sigs_by_name
-                        .entry(f.name.as_str())
-                        .or_default()
-                        .push(format!("{}({})", f.name, ptys.join(",")));
-                }
-                for (name, sigs) in sigs_by_name {
-                    if sigs.iter().any(|s| colliding_sigs.contains(s.as_str()))
-                        && sigs.iter().all(|s| colliding_sigs.contains(s.as_str()))
-                    {
-                        colliding_in.entry(i).or_default().insert(name.to_string());
-                    }
-                }
-            }
-            let (mut written, mut stubbed) = (0usize, 0usize);
-            for (mi, m) in mods.iter().enumerate() {
-                let mut src = gore_as::cache::emit::emit_module(m, &refs);
-                if let Some(names) = colliding_in.get(&mi) {
-                    for name in names {
-                        src = rename_free_fn(&src, name, &format!("{name}_g{mi}"));
-                    }
-                }
-                if src.contains("not fully recovered") {
-                    stubbed += 1;
-                }
-                let rel = if m.file.is_empty() { format!("{}.as", m.name) } else { m.file.clone() };
-                let rel = rel.replace('\\', "/");
-                // A cache entry's relative filename is untrusted: reject `..`, absolute, or
-                // drive-prefixed components so output can never escape `outdir`.
-                if std::path::Path::new(&rel).components().any(|c| {
-                    use std::path::Component::*;
-                    matches!(c, ParentDir | RootDir | Prefix(_))
-                }) {
-                    eprintln!("skipping {}: unsafe output path {rel:?}", m.name);
-                    continue;
-                }
-                // The lexical check above stops `..`/absolute paths, but a pre-existing
-                // symlinked component under outdir would still be FOLLOWED by create_dir_all /
-                // write, escaping outdir. Reject any existing symlink along outdir -> path.
-                let mut cur = outdir.clone();
-                let mut symlinked = None;
-                for comp in std::path::Path::new(&rel).components() {
-                    cur.push(comp);
-                    if std::fs::symlink_metadata(&cur).is_ok_and(|md| md.file_type().is_symlink()) {
-                        symlinked = Some(cur.clone());
-                        break;
-                    }
-                }
-                if let Some(link) = symlinked {
-                    eprintln!("skipping {}: symlinked path component {}", m.name, link.display());
-                    continue;
-                }
-                let path = outdir.join(&rel);
-                if let Some(p) = path.parent() {
-                    std::fs::create_dir_all(p)
-                        .with_context(|| format!("creating {}", p.display()))?;
-                }
-                std::fs::write(&path, src).with_context(|| format!("writing {}", path.display()))?;
-                written += 1;
-            }
-            eprintln!("emitted {written} modules to {} ({stubbed} contain a stubbed function)", outdir.display());
+            let stats = gore_as::cache::emit_all::emit_all_tree(&mods, &refs, &outdir)
+                .with_context(|| format!("emitting to {}", outdir.display()))?;
+            eprintln!("emitted {} modules to {} ({} contain a stubbed function)",
+                stats.written, outdir.display(), stats.stubbed);
         }
         AsCmd::Emit { file, needle, max } => {
             let bytes = std::fs::read(&file).with_context(|| format!("reading {}", file.display()))?;
